pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "discordplex"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "py-cord[voice]>=2.6.0",
    "PyNaCl>=1.5.0",
    "aiohttp>=3.9.0",
    "sphn>=0.1.4,<0.2",
    "scipy>=1.10.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0",
]

[project.scripts]
discordplex = "discordplex.__main__:main"
